# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: AutoViva AI API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ── Projects ──────────────────────────────────────────────────────────────
...
  /projects/{projectId}/credentials:
    get:
      operationId: listCredentials
      tags: [credentials]
      summary: List project credentials
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Credentials list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Credential"
    post:
      operationId: createCredential
      tags: [credentials]
      summary: Add credential
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CredentialInput"
      responses:
        "201":
          description: Credential created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Credential"
  /projects/{projectId}/credentials/{credentialId}:
    delete:
      operationId: deleteCredential
      tags: [credentials]
      summary: Delete credential
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
        - name: credentialId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # ── Test Suites ───────────────────────────────────────────────────────────
...
  /analytics/projects/{projectId}:
    get:
      operationId: getProjectAnalytics
      tags: [analytics]
      summary: Get analytics for a specific project
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Project analytics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectAnalytics"
  /analytics/score-trend:
    get:
      operationId: getScoreTrend
      tags: [analytics]
      summary: Get score trend over time
      parameters:
        - name: days
          in: query
          schema:
            type: integer
            default: 30
      responses:
        "200":
          description: Score trend data
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ScoreTrendPoint"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]

    # ── Auth ────────────────────────────────────────────────────────────────
    RegisterInput:
      type: object
      required: [name, email, password, role]
      properties:
        name:
          type: string
          minLength: 2
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
        college:
          type: ["string", "null"]
        role:
          type: string
          enum: [student, faculty, judge]
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
    AuthResponse:
      type: object
      required: [token, user]
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
    User:
      type: object
      required: [id, name, email, role, plan, createdAt]
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
        college:
          type: ["string", "null"]
        role:
          type: string
          enum: [student, faculty, judge]
        plan:
          type: string
          enum: [free, pro, team]
        createdAt:
          type: string
          format: date-time

    # ── Projects ────────────────────────────────────────────────────────────
...
    Credential:
      type: object
      required: [id, projectId, role, username]
      properties:
        id:
          type: string
        projectId:
          type: string
        role:
          type: string
        username:
          type: string
    CredentialInput:
      type: object
      required: [role, username, password]
      properties:
        role:
          type: string
        username:
          type: string
        password:
          type: string

    # ── Test Suites ─────────────────────────────────────────────────────────
...
    ProjectAnalytics:
      type: object
      required: [projectId, totalRuns, bestScore, latestScore]
      properties:
        projectId:
          type: string
        totalRuns:
          type: integer
        bestScore:
          type: ["integer", "null"]
        latestScore:
          type: ["integer", "null"]
        avgScore:
          type: ["number", "null"]
        passRate:
          type: ["number", "null"]
    ScoreTrendPoint:
      type: object
      required: [date, score]
      properties:
        date:
          type: string
        score:
          type: ["integer", "null"]
        runId:
          type: string
        projectName:
          type: ["string", "null"]
...