# lib/api-zod/src/generated/types/scoreTrendPoint.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AutoViva AI API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ScoreTrendPoint {
  date: string;
  /** @nullable */
  score: number | null;
  runId?: string;
  /** @nullable */
  projectName?: string | null;
}