# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AutoViva AI API specification
 * OpenAPI spec version: 0.1.0
 */
import type { UserPlan } from './userPlan';
import type { UserRole } from './userRole';

export interface User {
  id: string;
  name: string;
  email: string;
  /** @nullable */
  college?: string | null;
  role: UserRole;
  plan: UserPlan;
  createdAt: Date;
}